Allow build to work in unified OO build (#195)
The logic in lib/version.rb will look to see if there is `../pom.xml` and verifiy
it's the unified build and if so it will take the version/location of the inductor
JAR from the one that was just build. The default behavior is intended to be preserved
so that it doesn't interfere with the existing process.

There is still the issue that the gem itself is hardcoded to using the same version, but
we can fix this in another step.