Re-factored exec-order.rb
All changes can be split into 3 categories:
1) a little bit of re-factoring and code cleaning, put some of the logic into
library files, under its own directory - exec-gems
2) On every execution check that rubygems_proxy ENV variable matches rubygems
source on the VM. Fix if not.
3) On every execution check that all the gems listed in exec-gems.yaml are
actually installed on the VM. If not, re-create Gemfile and run bundle update.

Change #2 solves the issue where updating cloud service with the new rubygems
proxy requires touching compute component, to get the change applied to the VM.

Change #3 solves the issue where we may add or bump up a gem version in
exec-gems.yaml file, and have code that relies on the new gem/version,
but the VM wouldn't have it installed.