Work around for the older ruby in azure VMs
The issue presents itself when a component update is run and gem sync is found to be needed,
which for instance happens for Azure VMs, which now require fog-azure-rm gem.
The the problem with the older CentOS is that their version of ruby is not compatible with
the fog-azure-rm gem.
This fix restricts the scope of the issue to the volume and objectstore component, the ones
that actually need the fog-azure-rm.
For other components the gem sync will not run.