Changed how we decide to run initial bundle install
Before we would check for existance of Gemfile.lock, but that could lead to
issues when a previous deployment breaks during bundle install, because it
would result in deleting Gemfile.lock without creating a new one. So the next
run of exec-order.rb will see missing Gemfile.lock and will try to do bundle
install even if it's not actually required.
Now we will be checking for existance of the provisioner gem itself (chef,
puppet), and if the gem is not found we will run bundle install regardless