This PR will prepare exec-order.rb to perform update to existing system where we have older version of Ruby (e.g. 1.8.7).  This will also will be compatible with upcoming Fast Image effort where Ruby 2.0.0 is pre-installed.
This update procedure will not activate unless the cloud config
for the compute has been configured with env RUBY2_BINARY.  This
env must be a valid Ruby 2.0.0-p648 binary (pre-compiled).  This
also mean that it must be installed at the following location:
/home/oneops/ruby/ruby-2.0.0-p648.

On the first run when it is successfully acquired WRITE lock
only then Ruby update will happen.  This to ensure that we are
perform Ruby update with multiple proccesses.  Once the system
has been update exec-order.rb will be configured to use the new
Ruby.  At that point it will be just a wrapper to relaunch itself
by using newer Ruby.

Expectation on the inital run can run anywhere between 1-3 minutes
while Ruby 2.0 is being download and install.  Any subsequent runs
it will be normal.