Made sure vendor/cache folder on VM has correct owner. (#459)
Any file (gem) we add in vendor/cache during exec-order run will belong to root
because we run exec-order.rb as sudo. But the consecutive rsync command ran by
inductor is ssh-ing to the vm as oneops user and fails due to insufficient
permissions.
Fixed by explicitly changing owner of the vendor folder on the VM to oneops.