Reverted fast-image related changes in exec-order.rb (#57)
Even if we process a VM spun up from a fast image we still want to make sure
the gem sources and the gems on the VM are up-to-date. Otherwise, the changes
to gem sources and gemfile that we do from time to time will not be replicated
to the vm and deployments may start failing.