Fixed Vagrant Setup Circuit Sync Issue
Summary: The vagrant setup was being built using the open source code which has not been updated in months. It was causing the deployments to fail. Changes have been made to clone the latest code from the private repository and then sync all the clouds, models and packs when the vagrant is provisioned for the first time.